Up the version of the database
We should have done this every time the schema changed, and written
appropriate upgrade SQL. Unfortunately there are now devices in the
wild that initially created the database using older SQL. As such
they are missing columns. This causes them to crash when queries are
run against the DB expecting those columns to exist. As we didn't
up the version number, flashing a new build didn't cause the CREATE
SQL to run again.

Test: manual
Change-Id: I1daa263d9baa69af4abb4bebfe45586b1f9dfd0e

Former-commit-id: bdc407d11fdb98612f35a4ac52393be82dc04854