IntentForwarderActivity visible to instant apps
Prior to this change, attempts to resolve to an app in a managed profile
would result in a PackageManagerService crash as it replaced the
resolved results with the forwarding activity, which wasn't visible to
the calling instant app. This would result in a null ActivityInfo when
querying getActivityInfo. This change makes the forward activity visible
to instant apps so that in this case, the resolve info can be properly
populated.

Change-Id: Ib1336e561f966bde7a1f429ad71bae2e2c69bb5e
Fixes: 67489723
Test: manual - query and start from instant app in personal and work profile

Former-commit-id: 33ea16b61526f12c6c4f4807c907eb52c7fcb7f4