Add System API for system update clients
Introduces the concept of installation option and new system API for system
update clients to convert a complex SystemUpdatePolicy into a simple action
and an effective time, under any given time. This should abstract away the
current complexity in SystemUpdatePolicy and all system update clients only
need to query the current installation option and act on it, without the need
to understand or implement other logic related to the policy.

Test: gts-tradefed run gts-dev --module GtsGmscoreHostTestCases --test com.google.android.gts.devicepolicy.DeviceOwnerTest#testSystemUpdatePolicy --ignore-business-logic-failure
Test: runtest frameworks-services -c com.android.server.devicepolicy.SystemUpdatePolicyTest
Bug: 72939648
Bug: 64813061
Change-Id: I92537c7b99103a807d8adcbe2dc01b42a9740068

Former-commit-id: 658d7bbf1d5bedd435d8a90f7b9d841dc7cc513e