Do not allocate a hardware layer for TileLayout.
This was a regression introduced by ag/3429142 which replaced the usage
of PagedTileLayout with TileLayout in QS.
PagedTileLayout#hasOverlappingRendering was overridden to return
false, but TileLayout did not have the same optimization. Since QSAnimator
animates the alpha of this layout, a hardware layer was allocated
unnecessarily, potentially causing jank in the QQS <-> QS animation.

Bug: 72632226
Test: visual and using "Debug GPU overdraw" and "Show hardware layers updates"
Change-Id: Idfa8eeac0b63ae03942cccd59b52657ab263427f

Former-commit-id: 7e6e64c2156cc71586cd2474c56fea7a7bfba435