Use the WS hidden flag instead of WSA for dim check
The current code uses the WSA mLastHidden flag to ensure that the WSA
animation starts at the same time as the dim. However, the flag is set
to true when the layer is being preserved which will cause the dim to
hide and then show again. Instead, use the WS mHidden flag which is set
to false when the WSA will be shown, but still remains false when the
WSA surface is getting preserved.

Change-Id: I8b2545480fad78cdba26173d28503fc7fe591d28
Fixes: 72869695
Test: Bug no longer occurs.
Test: Other dims still work properly
Test: atest WindowManagerSmokeTest

Former-commit-id: ebcbc34587e0e88ddce2158d161c6eb39b06c3dd