Fix a11ynodeinfo parceling
The pane title was not parceled and unparceled symmetrically.
Fixing that.

Also, the CTS test that should have caught this error is broken.
Providing a test API to do parceling without recycling, so
parceling can be tested in isolation.

Bug: 73084962

Test: Updated CTS AccessibilityNodeInfoTest used to pass. After
the fix to parcel in isolation, the test failed. Fixing the pane
title parceling allows the test to pass again.

Change-Id: I161ce93f8bdea5d5dd2d28cd38b58633f7742e56

Former-commit-id: 6307240b8f50ddc5de878723964d766a8dabf6ea