Implement atom puller for ProcessMemoryState.
We would like to gather memory metrics of the processes when LMKD kills
a process. By gathering this info we would be able to analyze and improve
system health by potentially reducing memory footprint of the processes
contributing to the memory pressure on the device.

This feature would be available on production builds.

Perf analysis:
On an average getMemoryStateForProcesses method call takes
0.59 milli seconds per process.

To know more about this see: http://go/android-p-memory-metrics

Bug: 72177881
Test: Tested manually
Change-Id: Ice94c2d937a5c84bc205981d99588a30f686ff9b

Former-commit-id: 22d92b76bccd6666ea34009310dced62ba79165b