Don't propagate unrequested permissions
When an instant app is re-installed, we attempt to propagate the
previously accepted permissions. Instead of crashing the system_server,
we simply ignore the previously granted permission and forget that
the user ever granted it. If a second update to the instant app
re-requests the permission, the user will again have to accept or
deny.

Test: atest GtsPackageManagerHostTestCases:PackageManagerHostTest
Change-Id: I7bb9fb67159f6c4d4f4e2a24e7913d11713e826e
Fixes: 72698203

Former-commit-id: 1a3216bbb9faaa8ddd402b80e173c55f784706d7