Simplify volume dialog title for a11y
The dialog was asked to essentially self-voice its
transitions a couple of years back. This isn't consistent
with our general guidelines for a11y, and isn't easy to
support. Rather than updating the messages, which will
still leave this dialog inconsistent with others, I'm
just deleting the extra code.

Just setting the dialog's title to the correct string is
enough for GAR, so I've added that back in.

Bug: 73170352
Test: Confirming with TestBack that the correct title is
sent to accessibility services when the window appears.

Change-Id: If219f86e43fa262604a09226a02bdd4c08ae4c1a

Former-commit-id: 4ed5e98f82dc008200f25475ab39ceb5ac923fd0