Allow setting alarm to the same time as an earlier one
With alarm deferring, we may end up rescheduling an alarm to the exact
time we set the previous kernel alarm to. Removing the optimization to
prevent duplicate sets, as otherwise we may end up being in a bad state
of no kernel alarm set. Also added the field maxWhenElapsed to the alarm
dump for more details in the bugreports.

Test: Builds. Boots. Existing tests pass:
atest CtsAlarmManagerTestCases

Bug: 73069486
Change-Id: Iade4549b399def57902b644b1d50945722e60c28

Former-commit-id: 2fa9545e29ef6b6fc511c2136b5f91e5b7e29818