Fixes go to web behavior for Instant Apps
SysUI launches the installer with an ACTION_VIEW, CATEGORY_BROWSABLE
intent that has no data defined to indicate that the user has chosen
to go to the browser. A recent change to be more strict about what
type of intent is allowed to launch the installer broke this behavior
and is fixed by this change. We allow the special case in addition to
the install and resolve actions.

Bug: 73007515
Test: manual - launched instant app and tapped go to web
Change-Id: Ifc386018db3c7a6a0ed0a045adb972f0f8d9c667

Former-commit-id: 531db46112ac067efbc509910d1578d7b8a9d119