Expose getInstallerPackage to instant apps
Instant apps can now call this method and _may_ get a non-null
response. The return value could be null if the requested application
is not exposed to instant apps.

NOTE: It's possible for this to leak the installer package name. The
installer may not be exposed to instant applications and may be off
limits for other package manager query APIs.

Change-Id: Ib914bb9a946aad32b42dcbc0543d74007a28c76a
Fixes: 68277134
Test: atest CtsAppSecurityHostTestCases:EphemeralTest#testEphemeralGetInstaller01,testEphemeralGetInstaller02,testEphemeralGetInstaller03

Former-commit-id: cca0dc316cd67fba59721f2801d4329bc098e0fd