Remove mResizedWhileNotDragResizing
This codepath is fundamentally broken with the hierarchy model. It's intended
operation is such: Preserve the WindowStateAnimator level surface freezing
updates to it, while waiting for a new surface to draw at a new size.
Once the new surface has drawn, show it at a new position while hiding
the preserved surface. Via this mechanism it could achieve atomic updates
of size and position. However position is now handled at the WindowState level
and so preserving the WindowStateAnimator surface doesn't achieve the intended
affect of freezing the position. Furthermore the new location for setting
the flag (WindowState#onResized) seems to have issues as it is now
being triggered during resize. This is triggering Surface preservation
during rotation making seamless rotation impossible. It also causes issues
in other resize scenarios (drag resizing, PiP, virtual display (b/72220802)) and
so we have to disable it bit by bit. We still have some need for atomic
update of size and position but the solution is unlikely to resemble this
code and so it seems best to remove it for now.

Bug: 72038766
Bug: 72220802
Test: Manual. go/wm-smoke
Change-Id: I6f2f5f6b26957ffa49c0bbaae5e099abf13ac47a

Former-commit-id: 09286c9e30d2781a5687c422bbe215e09851f10c