Ensure nav bar contrast under magnification
When the nav bar is transparent, the app background
provides adequate contrast with the buttons (at least,
it seems to). But magnification zooms in on the app
window but not the nav bar, which can bring a section
of the UI behind the nav bar that is the same color
as the nav bar buttons. That makes the nav bar quite
difficult to see.

This change makes PhoneWindowManager listen for
magnification updates. When magnification is zooming
in, it makes the nav bar opaque.

The callbacks happen when magnification goes from 1x to
not 1x, and from not 1x to 1x. So we only call back when
the nav bar needs to change, not for every movement made
by magnification.

Bug: 36677486
Test: Enable magnification, zoom in, and watch the nav
bar become opaque. Watch it become transparent again
when we zoom out.

Change-Id: Id9b2c57af98f6cefc81ecf3d7bdab4e49ac07474

Former-commit-id: d7551607b34cb15725200a3de24cc2feb20f1b44