Move performSurfacePlacement after tryStartExitingAnimation
In performSurfacePlacement, there's downstream code to update
the navbar color based on whether the IME is visible. However, IME
is only marked not visible once tryStartExitingAnimation is called.
Therefore, tryStartExitingAnimation needs to happen before the
attempt to update the navbar color so it can know that the IME
is going to be hidden.

Change-Id: I3a337d1b3a2caea8694204111f778924d2fd3ec0
Fixes: 71877162
Test: After closing IME in dialer, navbar color returns to original.
Test: go/wm-smoke

Former-commit-id: abedd3226318792018996b8a3f35f794a72dba73