Prevent freezing the display while rotating seamlessly.
Following recent changes we are now receiving WindowManagerService#reconfigureDisplayLocked
during rotation. This is interpreted as another change in display
configuration that requires screen freezing and we continue on to startFreezingDisplayLocked.
This triggers the screenshot based freezing, but we were already trying to seamlessly rotate.
It seems a sensible solution to just disable screenshot based freezing while seamlessly as
we now disable screenshot based freezing while already using it.

Bug: 72038766
Test: Manual. go/wm-smoke.
Change-Id: I705518be11302cfaa95207e3b3d0c7e93e54ebbf

Former-commit-id: efc7570b780819a3095867a32b41dc07292ec87b