Create display if necessary when adding windows.
It is possible for a window to be added to a display before
ActivityManagerService has processed a display creation. In this
case, the relevant WindowManager and ActivityManager containers have
not been created. This will prevent the client from proceeding.

This changelist addresses this problem by allowing the WindowManager
DisplayContent to be created in this specific scenario. When the
ActivityManager side processes the new display, it will be associated
with the existing DisplayContent.

Change-Id: I1bc7d11f5cc538f35b5f7b0a8303d48fba370ea0
Fixes: 73292997
Test: cts-tradefed run singleCommand cts-dev --module CtsMediaTestCases --test android.media.cts.EncodeVirtualDisplayWithCompositionTest

Former-commit-id: 1cddfdc103fcfc6a8652249d3f7f04d6ffc4766e