Fixes for inflating AnimatedImageDrawable
Bug: 73529437
Test: I7907f5dd7eb8d4ab72b8e9ddcbcbfc7aa8cb05ae

Support AutoMirroring, from inflation and setting manually.

Ensure that AnimatedImageDrawable always has a State object, even if
it has no mNativePtr. If it has no mNativePtr, throw an Exception when
trying to use it.

Require that inflation have a valid src, unless it is waiting on a theme
(like BitmapDrawable).

Change-Id: I5b22cdbf4b57862d6ccc5750a677168287860f9f

Former-commit-id: 7f66873e0edfb9a4ad6b463c2839508f3453ad57