WebView: fix docs for setSafeBrowsingWhitelist
No change to logic, documentation change only.

This changes "URL" to "host" everywhere in this API documentation.
This API never accepted a list of URLs, it only ever accepted a list of
hosts (e.g. "www.google.com" is ok, "http://www.google.com/" is not).
The original documentation was misleading.

Test: make docs (manually verify)
Change-Id: I3b57a2a160ff00ba8f92b30f6e3b4399e55f2ea0

Former-commit-id: 5a6ae8c739419fb5ea445e791d28d4674f5dade9