Cherrypick "Add device config to decide which Auto Selection Network UI to use."
This change added the config because the HAL V_1_2 only supports Pixel 3, and the new Auto
Selection Network UI is based on HAL V_1_2. So we set the flag to decide which Auto Selection
Network UI should be used based in the device type.

Bug: 63718613
Test: Basic telephony sanity
Merged-In: I6b54db63b318564983ae11cffaf7fc27df71b060

Former-commit-id: 08941a5f3efd7def6e840fc04ae407f64be7adc1