Support Slices in Settings Search
Add an API for Settings search to associate Uris to Keys,
to enable support for Slices in search.

We add a separate API rather than updating IndexableData because:
- Apps will need to support a method of collecting all slices for the
Slices API, and that method can be shared between the Slices API and
this one.
- To update search indexableresource, we would need to establish a
contract on how to declare a slice Uri in each xml. Even if we did that,
the providing app would need their own logic to parse the Xml for
the Slices api, so we can't save them work.

Bug: 62807132
Test: contract files
Change-Id: Ie8853ac57643231794afdbb2f7ce1cc0f77619a2

Former-commit-id: 51598602e39570ef482d830eb013af44793852d5