NetworkCapabilities needs UID to be satisfied.
A recently added NetworkCapabilities feature expects that a
NetworkRequest defines a real UID in order for it to be satisfied
by a VPN-based network.

So when creating our JobStatus inside the system (either from a new
external request, or from a persisted job on disk), splice in the
source UID of the job, so that VPNs start matching.

Test: bit CtsJobSchedulerTestCases:*
Bug: 72765718, 73418534
Change-Id: Id611a98fee8f53cea92f7a22ec02f7f6d8a4c534

Former-commit-id: fb0f97171f14520af1b002b9227b25eb8980bd27