Restore "Add "Unlocked device required" parameter to keys"
Add a keymaster parameter for keys that should be inaccessible when
the device screen is locked. "Locked" here is a state where the device
can be used or accessed without any further trust factor such as a
PIN, password, fingerprint, or trusted face or voice.

This parameter is added to the Java keystore interface for key
creation and import, as well as enums specified by and for the native
keystore process.

This reverts commit d5a34329ae0ffc08056b924c6cf68a94bb3c63ab [formerly da82e2cb7193032867f86b996467bcd117545616].

Test: CTS tests in I8a5affd1eaed176756175158e3057e44934fffed

Bug: 67752510

Change-Id: Ia162f1db81d050f64995d0360f714e79033ea8a5

Former-commit-id: d7c961ee914192e09ec10727da6d31a6b597bf51