Add alarm info to QS Panel
Added alarm info in the same space as the tooltip. Moved the header info
out to the QuickStatusBarHeader to consolidate all header views.

Since, theoretically, there can be some weird races between the
animations of the long press tooltip and the alarm info, added
protective animation state code in the header (such as fading out the
long press tooltip before animating in the alarm text).

Fully removed the alarm tile.

Test: Visually/manually
Bug: 73764084
Change-Id: Ic1ec450e560a56567f9efeb04d3efbd916f05bc3

Former-commit-id: d3cf756aa41c9385528651ffe6da7faad6b28fbc