Add columns for snapshot table
Currently snapshots are held in memory, meaning they must be regenerated
if a reboot occurs before a key sync. Also, when debugging, it is
difficult to know what version of e.g. the server params was associated
with a particular snapshot, as this can be mutated after the snapshot is
generated. This change adds the required columns to the DB contract for
storing snapshots. In subsequent CLs the update SQL will be added.

Test: none, no functionality added
Change-Id: Ica866b06950a5801e8a2c3641e79706bbbf48384

Former-commit-id: 5dd87d8827fcfa22ee7a1973c8e73d7354752cf7