Wait to pause activity if translucent activity is waiting.
A recent change moved an activity to paused if it was stopped and
became visible. However, there are situations where this is not
desired, such as when waiting for translucent activity to become
visible. The extra binder calls from pausing during this time
leads to visible jank. Therefore should avoid moving states until
after this operation has completed.

Change-Id: I7bbab002cbee479f436515e7cc5db6f3a1f10b16
Fixes: 73832787
Test: atest FrameworksServicesTests:com.android.server.am.ActivityRecordTests#testPausingWhenVisibleFromStopped
Test: atest CtsActivityManagerDeviceTestCases:ActivityLifecycleTests#testPausedWithTranslucentOnTop
Test: atest CtsActivityManagerDeviceTestCases:ActivityLifecycleTests#testPausedWhenReturningWithTranslucentOnTop
Test: atest CtsActivityManagerDeviceTestCases:ActivityLifecycleTests#testPausedWhenRecreatedFromInNonFocusedStack
Test: atest CtsActivityManagerDeviceTestCases:ActivityLifecycleTests#testPausedWhenRestartedFromInNonFocusedStack

Former-commit-id: a0fb8e03b04759fca82428e7ad70c2832cb94d73