Assign leash layer when starting animations.
Currently AppWindowToken does this but WindowContainer does not
ending in a situation where WindowState level animations can cause
child windows to be ordered below the IME when they would otherwise
be above. I couldn't find any reason for the assignLayer call that
was here, as the parent won't view the leash as a child from the WindowContainer
perspective and thus assignment will never happen. I've just replaced
the old call with the new layer assignment.

Bug: 72173720
Test: Manual. go/wm-smoke
Change-Id: Iba57158645e27c2ff74ce6e0b59cd15b15bfe3f5

Former-commit-id: f12f9d3ded0d947b1b0c9fd71cf75bdb83f5005a