Do not call checkReadyForSleep recursively.
resumeTopActivityUncheckedLocked can be recusively called. There are
currently measures in place to guard against
resumeTopActivityInnerLocked from being called inside recursively.
However, calls to checkReadyForSleep in the same method are not
protected in the same way. This can lead to a stack overflow.

This changelist moves the call to inside the scope protected from
recusive invocations.

Change-Id: Id5a7e04a354032ad10cf285f8d6b26d666734e6e
Fixes: 66344012
Test: go/wm-smoke-auto

Former-commit-id: fc7cedd093c97600b3b5cc73361db7ce3b571d2f