Creating new user correctly connects to the system OverviewProxy
When creating a new user, it creates a new OverviewServiceProxy which
connects to launcher instead of the system user version. Since the new
proxy does not have any callbacks connected to it, launcher cannot send
any data to change systemui. Preventing the new proxy from connecting
allows the original to connect.

Test: manual, create new user, wait for connection, play with quick step
Change-Id: I0838fe8294c7a3ee89062c0c9855caf5ab52482b
Fixes: 74001992

Former-commit-id: 1b1d346b3547809f410a62ec2cdbd3a6ff21127a