Increase rotate suggestion timeout in fullscreen
To use the rotate nav button in fullscreen the navbar must
become visible. This normally requires a user gesture which
makes it hard to hit the button given its normal timeout
duration. Make it easier to tap by effectively extending the
timeout if the suggestion is given when the navbar is
invisible.

Test: manual with fullscreen testing app
Change-Id: I54ac3dfb590df9b9a039c53984623497ca18fcb5

Former-commit-id: 5aeca79673d142f9924f71fc6055c46d837a95e7