Fixes KeyButtonView to dismiss ripple after aborting for long press
When set pressed state to true after aborting gesture to show ripple,
the set pressed state was not set back to false even though it receives
a cancel motion event. This leads to showing the ripple without
dismissing it. To dismiss the ripple, when receiving the cancel motion
event, set pressed state to false.

Change-Id: Ic3b9c8e1ecd6cff8ba7a34c2c179a9810f962862
Fixes: 74136297
Test: With an app open in fullscreen, long press home

Former-commit-id: 6af4d0965d09c8b918109b4e939681b0f90d3544