Multi Cutout: Fix more cutout issues
Ensures that the navigation bar extends all the way to the bottom of the display.

Also ensures the nav bar window is properly padded.

Also ensures that windows that do not hide the navigation bar can extend behind the cutout if
the cutout is within the navigation bar region.

Also ensures that wallpapers (just like dreams) are not letterboxed

Bug: 74195186
Test: Turn on multi-cutout; verify the wallpaper and navigation bars extend all the way to the bottom
Test: Open an immersive app, verify it gets letterboxed
Change-Id: I724080edc26d7ee09dba690f6670d4f66d1c93f6

Former-commit-id: 7e58a08ccdd528b48497d7286be3c28cc0e339c0