Encode F16 Bitmaps to 8-bit P3
Bug: 73743812
Bug: 71430152
Test: If2c3ee0f32eff77b11ce5a7fe82c02811ed1beb3

This matches how we handle decoding images with a wide gamut. For WebP
and JPEG, we cannot encode with more precision anyway. For PNG, we could
in theory encode to 16 bit linear PNGs. This would be more precise, but
would still throw away information due to normalization. Using 8-bit P3
matches the behavior for the other formats, so it is nice to be
consistent.

Change-Id: I69a44a3eff70e26448a75ecc63ed7604b9c74a95

Former-commit-id: 4856210398468fd5196e1cb68b80e7529e1046a0