Moved urlBarResourceId from autofill service manifest to whitelist settings.
The manifest attribute is still public as it might have been used by autofill
services deployed against P DP1; it will be removed after the next developer
previs is branched out. We also need to assumie a default value for the buttons
if not specified by settings, but that will be done in a separate change so it
can be easily reverted.

Also implemented support for multiple buttons, and added unit tests.

Test: atest CtsAutoFillServiceTestCases:VirtualContainerActivityCompatModeTest \
CtsAutoFillServiceTestCases:VirtualContainerActivityTest \
FrameworksServicesTests:AutofillManagerServiceTest

Bug: 74445943
Bug: 72811561
Fixes: 73786629

Change-Id: I066ecf40fde2c5318dd8633a659fca8b7af8aecd

Former-commit-id: 1c7eb0c8eb156c4656f2d3e1b97929797e4e7ec9