Use isKeyguardLocked to determine whether to resume.
Previously, we used isKeyguardActive to determine whether we should
resume an activity. This does not account for the keyguard going
away, which can leads to a timing issue where we do not resume an
activity when coming out of sleep from fingerprint unlock.

This change also prevents us from pausing a stopped activity that is
is considered to be the resumed activity as it will be brought to
the resumed state.

Fixes: 73991496
Test: Open display settings, lock screen, fingerprint unlock, press
      back, verify main settings page is shown immediately.

Change-Id: I3a7a33f5359a41fc958e3d8e799c39f5193ecb56

Former-commit-id: d939cf04e52b4f8e635989374e3bc47c3936aedc