Show dynamic subtext for RestrictedSwitchPreference.
Show dynamic subtext for RestrictedSwitchPreference when corresponding
restriction is on according to the current settings status controlled
by this RestrictedSwitchPreference. I.e., show "enabled by admin"
when corresponding setting is currently turned on by admin, or
"disabled by admin". In the past, this string is hard coded in xml and
the subtext of RestrictedSwitchPreference always shows one of
"enabled by admin" and "disabled by admin" regardless of the current
status of the setting.

Bug: 73811933
Test: manually by TestDPC
Change-Id: I2bbeee14e650b288d71f6aa59fc63dd78e18d49e

Former-commit-id: 1177b369e5726056e52f8f914d4c63613b8ad19a