Update comment for HAPTIC_FEEDBACK_ENABLED setting
The current comment is a little confusing, and this ends up in the
public documentation here:

https://developer.android.com/reference/android/provider/Settings.System.html#HAPTIC_FEEDBACK_ENABLED

See b/22390263 entry for a link to at least one instance where an actual
developer was confused by this.

This CL changes the comment to match what we call the setting in the UI
of the Settings app. Note that it's possible this could become stale
someday (e.g. we used to label this "Vibrate on touch" and switched to
"Vibrate on tap" at some point), but this description will probably
still be ok and remain an improvement over what we have now.

Bug: 22390263
Test: N/A (comment only change)
Change-Id: I717a7c5a2f9ecc38cfe6f0c1c0379a868f810782

Former-commit-id: 4457bb8e0427f05f8c747ef7d106b169f338e4e2