Make the GnssStatus constructor public and mark as hidden.
Android Things user-driver model for GNSS requires that the GNSS driver
to be implemented by the developer in Java and have its updates
forwarded down to the GNSS framework. This means that the developer will
need to be able to create GnssStatus objects.

A GnssStatusBuilder class is being added in ag/3726284, intended to be
used to create GnssStatus objects. We do not intend for the developer to
construct a GnssStatus object directly, so while this constructor is
being made public, it's also being marked as hidden.

Test: Builds.
Change-Id: Ie8e75e2bc264fe8b1981afebb3f4d10fadd89915

Former-commit-id: 9c617f5677541fd3fd5fc2a6f74b3772b86d54d5