Fixed data connection state callback not working
The old behavior of onDataConnectionStateChanged callback is that
if at least one data connection for one APN type (e.g. default,
mms, supl, ims, etc...) is connected, then it reports connected.
It only reports disconnected if data for all APN types are
disconnected.

Since now we have data connections are always connected. For example,
even if the user disables data, the unmetered IMS data remains
connected. In this case the client's onDataConnectionStateChanged
won't be called.

Now we changed the behavior of onDataConnectionStateChanged to
report the connection state for default APN type only. This matches
the behavior of TelephonyManager.getDataState().

Test: Manual
Bug: 73226433
Change-Id: I0744cf33337864c946af6e7d186bab78584acfa4

Former-commit-id: b1bac54d609703ec5fbc633fab89c73994289496