Create snapshot even if no PendingIntent is registered
When the user first unlocks the phone after booting, the system
app has not yet started. As such, it will not have had a chance to
register a PendingIntent. But if it has ever previously initialized,
the framework can still create a snapshot, and should. Otherwise, it
may be up to 72 hours before the user unlocks their phone again using
the primary method, which adds delay to the key sync.

Bug: 73921897
Test: runtest frameworks-services -p \
      com.android.server.locksettings.recoverablekeystore

Change-Id: Idfaf53194e6a2f5d5ce0123d72f46197392d2c99

Former-commit-id: 2fd4b597ae3cfaa5dfa8156ec15bc813d69acf7a