Fix stuck exiting windows
When a window was marked as exiting because in WS.removeIfPossible
applyAnimationLocked returned true just because AWT was in
mClosing/OpeningApps for the transition, the window wasn't removed
if there was no animation running for the app transition.

Fix this by also calling the animation finished callback when not
running an animation, because other code parts rely on that being
called.

Test: Change language, swipe up, make sure no stuck window
Test: AppWindowTokenTests
Change-Id: I9cfd0c4090a258bc90ba5cf7d71795b6fce7f8ef
Fixes: 72811863

Former-commit-id: ab9fcb2197796ab80cbfb226c443e4eb3ffbc15b