Bump activity stop timeout past ANR timeout.
Once the activity stop timeout is triggered we may destroy
surfaces triggering native crashes. Imagine though an
app that hung in onStop, and then receives (and fails to handle)
an input event. Since the stop timeout was started before the hang
the stop timeout is very likely to trigger before the ANR timeout
leading to difficult to account for crashes. By the same logic we
bump the toast removal timeout.

Test: Feed it to the monkeys.
Bug: 62536731
Change-Id: I8e357271cfaa9f4077bb5bc92fdbd4584a425e8b

Former-commit-id: 3406d468d7a808a87871138f4ab2f9f9ff016c71