Enable cleartext for captive portal apps
When a user attempts to visit a captive portal the Activity will attempt
to load http://connectivitycheck.gstatic.com/generate_204 and a
walled-garden will typically intercept this and respond with a 302
redirect to a sign-in flow.

The new default of disabling cleartext unless explicitly enabled breaks
captive portal operation. This change enables cleartext for the two
applications that contain a CaptivePortalLoginActivity.

Bug: 75279991
Test: Manually associated with captive portals.
Change-Id: I081dfed067b3a78a81bfd3ec10afaf953b6baf3f

Former-commit-id: d816a2a8ed1cfc32cff37618418e706104b9a264