Statsd: remove DurationAnomalyTracker.resetStorage
The explicit resetStorage function in the subclass isn't needed.
It does exactly what its superclass does.
The previous check for !mAlarms.empty() was actually not needed, since
there can be legit cases where !mAlarms.empty() (namely, when the
mMostRecentBucket is moved forward initially - we may call reset, but we
still want the old alarms to remain).

Test: still pass
Change-Id: I3ffb6e97c02b75170c246f57d55ccf59cd368b9c
Fixes: 74607818

Former-commit-id: 6a1d3af7f5167be234cf8b77b927c6b366660bfa