DisplayCutout: Correctly apply light status bar
Fixes an issue where light navigation bar was not
properly applied in landscape where the letterbox
does not overlap with the status bar.

This removes a hack that is no longer necessary
now that we track exactly if the letterbox overlaps
the system bars and put them into black-background
mode if so.

Bug: 72696928
Test: Open API Demos > Activity > Max Aspect Ratio > Any; rotate to landscape, verify that status bar is light.
Change-Id: Ifc6355e2eeadf0821c9ac82ce78c812916de5034

Former-commit-id: 0c124c6ad79b1b7a481422e27853655d228eb2f8