Correct exiting pip animation state.
We were using mForceScaleUntil resize to persist
the pip animation state past the Stack animation state
to give the client time to request a resize. However it now
seems the client is almost always requesting a resize before the
stack animation state ends. To account for this we need to make sure
we only reset mForceScaleUntilResize once at the beginning
of the animation and then allow the flag to be cleared ahead
of the Stack animation state.

Test: Manual.
Bug: 69065873
Change-Id: I9b5260ac984270ed62e75a4369360c0843e34d28

Former-commit-id: 2025bf84da365fa32d1735bb309e0004b6b217f1