Distinguish between bringing stack to front and delivering to top.
When we are in split-screen mode and are delivering an intent an
activity in an unfocused task, we report back we are bringing the
task to front. This is sometimes incorrect as the task is already
visible.

This changelist addresses this issue by checking the launch stack's
position in split-screen. If it is the topmost stack of its type,
then we will report back START_DELIVERED_TO_TOP instead of
START_TASK_TO_FRONT.

Change-Id: Ic900a8195a40fd485551a90d1a66e055b0b2539b
Fixes: 72411159
Test: atest FrameworksServicesTests:ActivityStarterTests#testSplitScreenDeliverToTop
Test: atest FrameworksServicesTests:ActivityStarterTests#testSplitScreenTaskToFront

Former-commit-id: 32e09ef7862952586b4c69cf38e1a5d91ebce47d