Layer input method on top while transitioning.
When exiting an app, we need to keep the IME on top of it
for the duration of the closing animation. Previously
this was accomplished by having the app remain as the IME target
based on a comparison of mAnimLayer with the new layer. The new
architecture makes it difficult to do any such direct comparison.
We opt instead for the approach of promoting the IME above all apps
during app transitions.

Bug: 74427597
Test: Manual. go/wm-smoke
Change-Id: I0fb66b4cab8aa474c7a64d0cefd32310151d544e

Former-commit-id: 234b63388563c2336acdb3270aed5fab552f3e23