Rename persist.traced.enable to sys.traced.enable_override
Allowing for both a default in the system
image and an override form the server requires at least two properties.
Given that we're using two properties anyway we can change the
server override to match the pattern of the other flags and not be
a persisted setting.

Test: Manual.
Bug: b/76077784
Merged-In: I98932085430cdeaf4ccd036a0751f119c0cdeb3c
Change-Id: I98932085430cdeaf4ccd036a0751f119c0cdeb3c

Former-commit-id: c5f06ac35e9a1d7d209820c0543c24892e990949