Fixes the inability to start quickscrub because hit target is none
This can be caused by going home and quick scrubbing. onLayout in
NavigationBarView will record the incorrect location of the home button
if the home button has been dragged. The location is used to determined
which button was touched. Fixed by only recording the location if the
button has not moved.

Test: manual - press home, quickly scrub far and try scrub again
Change-Id: I24f1ac44531a030fdd4ba23410d803908462d564
Fixes: 75975208

Former-commit-id: 3b7e39014522cc1838a56a19fc9433f5b15adead