Make Assistant stack invisible behind split-screen stacks.
Assistant stack shouldn't be visible behind split-screen stacks. In addition
to this not making sense, it also works around an issue here we boost the
z-order of the assistant window surfaces in window manager whenever it is
visible which makes them appear above the split-screen stacks eventhough
they should be below.

Change-Id: I34220d5aec485ccfd89c9f202367ef8432a88a78
Fixes: 76106972
Test: atest FrameworksServicesTests:com.android.server.am.ActivityStackTests

Former-commit-id: 8e923af90f6b04b12bfe7c402c364c65c42e0bae