WM: Move mLayoutSeq to DisplayContent
Fixes an issue where the mLayoutSeq could advance due to laying
out a secondary display, but that would not advance the mLayoutSeq
of the WindowStates on the primary display.

Because the mLayoutSeq no longer matched the global value, the
windows on the primary display would then not get the resized
notification.

Fixed by making mLayoutSeq a property of the DisplayContent, such
that for different displays, mLayoutSeq is independent.

Change-Id: I9d6b95117e8428b0f2f5e88a9260ea6514074fe5
Fixes: 36781810
Test: atest WindowStateTests DisplayContentTests

Former-commit-id: 5251b1df577d99cd6b39ca21b1e2887c16462063