Fix flickering when switching display power modes
So far we've been Waiting for the next frame but it doesn't
mean that it will be visible on the screen.

We're usually rendering with 2 buffers and it's possible that
we'll switch the display power mode before anything has hit
the display.

Increased the timeout for now until we have a better alternative.

Change-Id: I63afc8a9139a4ccbba4f16f5569717e22217da9a
Fixes: 72527083
Test: manual
Test: packages/SystemUI/tests/src/com/android/systemui/statusbar/phone/ScrimControllerTest.java

Former-commit-id: 0791d97e117f4ebbb960852c1c5aa6c26de089b0