[RTT] Expose new public API for number of measurements
The RTT distance and standard deviation are calculated based on a set
of N measurements. Of those M (M<=N) may be successful.

The quality of the provided results can be assessed by 2 measures:

1. The standard deviation of the results

The standard deviation is only valid if the number of successful
measurements is >1 otherwise a 0 is returned. A value of 0 cannot be
differentiated from a legitimate 0 standard deviation.

In one case (number of samples > 1) a standard deviation of 0 indicates
high confidence in the result. In the other case (number of samples = 0)
no confidence information can be derived.

2. The ratio of successful to total attempted measurements

Add API which provides the number of attempted measurements and number
of successful measurements used to calculate the final results.

Bug: 76445930
Test: N/A (unhide CL only, code verified earlier)
Change-Id: Ia7322b3abc28e83c29d0b0ae79db30fd99bb7fe1

Former-commit-id: 7a60ea4c6a2e4426465acd0824cd9ca3a8b9e158