Use package manager stock flags to pull lib info
In several places activity manager components were not using
the stock package manager flags and as a result were not pulling
the libraries on which the app depended. This leads to subtle
cases were the dependent libs are not present on the classpath
resulting in app crashes.

This is a manual merge of I33cb4464599dd8c1b68f64eb3bb7c470bf27ab97 from
o-wear-dev

Test: Tried to repro original issue
Bug: 73760062
Change-Id: Ia5c9236b330a8595c2c23b39aa73f0a07a277154

Former-commit-id: cff8f0f4e7c1e0bbf2a5a5451db70d0e047e2560