Nav icons and quick step states are determined by service enabled
When the device quickly changes states (such as language or font size),
it would disconnect and then reconnect the service and as a result a
flash of the nav buttons will occur. Since the nav bar is connected to
launcher service unless 3rd party launchers are used, that should
determine if the icons should change or quick step should be
enabled/disabled instead of if there is a connection. Also hold the last
interaction flags and apply on startup until service is disabled (using
3rd party launcher).

Test: go to settings and change the font size under display
Change-Id: Ibebc8c3f8493574dd21774bd292c2096cfdad469
Fixes: 74199057
Fixes: 73483047
Fixes: 72225004

Former-commit-id: 8a7c97a16d7cfa7bcc602dfa969370bce9dfb686