Improve the JavaDoc for SOFT_INPUT_STATE_UNSPECIFIED
SOFT_INPUT_STATE_UNSPECIFIED is a tricky mode because its de facto
spec is that the system can do whatever it thinks valid.

Despite the fact that this is the default value for newly created
Windows, unfortunatelly there is no clear spec about what would happen
regarding whether the system thinks the software keyboard should be
shown or dismissed.

This CL doesn't change the behavior, but it just attempts to clarify
the behavior in JavaDoc.

Bug: 73793727
Bug: 77152727
Test: make -j doc-comment-check-doc
Change-Id: I69dab7c7018da89ba57fe2f5c8e718ef95985307

Former-commit-id: a754785cf949113a7ae09893c101558f4737f3cd