Update triggering logic for hybrid notification
Currently in the worst case a user can receive 4 battery
notifications if the time estimates and percentages line up right.
This CL makes it so a user can at MOST receive one "low" battery
warning and one "critical" battery warning per charge cycle. A
charge cycle is restarted when a user charges to at least 45%
battery AND has 6 hours remaining. This does not affect the
behavior of the non-hybrid notification.

Test: robotests
Bug: 76203825
Change-Id: Ib3c7fe589f1ce4c0cdb821e1f21d1139a56fad62

Former-commit-id: fd38aa5114daf9656e38f1a64faff7cb435e3049