Ensure ActivityStack's cached resumed activity is updated.
ActivityStack stores a reference to its resumed activity for quicker
future reference. Currently, children must inform the ActivityStack
of any movement in order to properly set this value. We are missing
cases today, causing stale values to persist.

This changelist addresses by updating the stack's paused and resumed
activity references whenever the ancestor stack changes when a task
or activity's parent is changed.

Change-Id: I5ef6db6af3d7e304012c54afa996fbfff28e9110
Fixes: 75992227
Test: atest FrameworksServicesTests:com.android.server.am.ActivityStackTests#testResumedActivityFromActivityReparenting
Test: atest FrameworksServicesTests:com.android.server.am.ActivityStackTests#testPrimarySplitScreenToFullscreenWhenMovedToBack
Test: atest CtsActivityManagerDeviceTestCases:ActivityManagerPinnedStackTests#testFinishPipActivityWithTaskOverlay

Former-commit-id: 84730a0a0c8e43877c5c4ba00231b4daf44489d0