[Notif] Disable animations as screen is powering off
When showing an incog notification, it's getting added to the remove
anims, but the screen is turning off and animations are never getting
run (which means the transient view gets added but it doesn't get
removed).

The root cause is notifications clear temporary views in the power on
(due to collapse animations), but don't clear them when unlocking via
fingerprint.

Fixed by clearing remove animation setup (where transient views are
used) when disabling animations. This will stop the transient views from
being held on post-screen off/when animations are disabled.

Test: Visually with logging (tested before & after)
Test: Dumped hierarchy and checked that transient views weren't being mismanaged
Fixes: 77217401
Change-Id: I48835e45fd76c5e74810f5256c14b0c4c2ca3ec4

Former-commit-id: 8ee5365f7f15f1cdf42cc3cfcd5d627ba673d580