Properly remove activity from task on start failure.
When we fail to start an activity, it is removed from its parent task
by the starter. We currently directly ask the task to remove the
child activity. This is incorrect as the stack is unaware of the
movement and cannot cleanup the task if necessary.

This changelist addresses the problem by asking the stack instead to
finish the activity.

Change-Id: I41faa2f4e81873601d45a8ccd1bf22fef9e57033
Fixes: 76679042
Test: atest FrameworksServicesTests:ActivityStarterTests#testTaskModeViolation

Former-commit-id: 2b8e0378ba7eda2203a7152a689c100259ae52ef