Deprecate InputMethodUtils#isSystemIme() part 1
There are not a small number of files that import InputMethodUtils
just to use InputMethodUtils#isSystemIme() is necessary.

In order to make InputMethodUtils purely an internal utility library,
it would make much more sense if the logic is implemented as a private
API of InputMethodInfo, like methods like
InputMethodInfo#isAuxiliaryIme(), rather than in a static method of
InputMethodUtils.

This is a purely mechanical refactoring.  There should be no behavior
change.

Bug: 77730201
Test: atest FrameworksCoreTests:com.android.internal.inputmethod.InputMethodUtilsTest
Change-Id: I9c8518988787b748ebb35fc86fe6beee1d6c633d

Former-commit-id: fd70fe8fe122560541eb4261357d8edc92dc55dc