Allow extra height for the smartreply buttons
Currently ExpandableNotificationRow assumes that notifications can be
up to @dimen/notification_max_height big and propagates that value to
NotificationContentView. However, when smart reply buttons are present
the content could get bigger than that and cause content to be
truncated in an undesirable way.

Add a smart_reply_button_max_height dimension and add it in addition
to notification_max_height when smart reply buttons are present.

Screenshots:
- before: https://screenshot.googleplex.com/MssT483YPd5.png
- after: https://screenshot.googleplex.com/k5StCMhb74t.png

Bug: 72360319
Bug: 73949609
Test: In Notify post a Messaging notification with Large chices
Change-Id: Ib3e43cff927bb5ca6dd941705cd280dc85f14a08

Former-commit-id: e7cf498342b225953f5071457ded0bb995656ffe