Fix IME target after rotation
When rotating the screen, and the top activity relaunches, it will
add a new window and make it visible. However, we never run an app
transition, so any of the calls to computeImeTarget in there don't
run. Initially, isVisibleOrAdding will return false because the
client adds it with INVISIBLE, and then relayouts it. However,
after adding it there will never be a call anymore to
computeImeTarget, so we add a separate call when a window becomes
visible.

This is a really targeted fix to a broken concept (the way how we
guess the IME target) that has to be revisited very soon.

Test: go/wm-smoke
Test: Open Gmail compose, open status bar, rotate screen, ensure
that IME ordering is correct in "dumpsys window visible".

Change-Id: I345b535cd531dbd8955e9a4cde2b87f0b7f41a88
Fixes: 75007549

Former-commit-id: 782d10375c989f21a66d3c91ccce2c28b993711c