Don't update crop WSA when drag resizing so screenshots are correct.
When drag resizing, there's no need to recalculate the crop for
the WSA since it will not be affected by the decor frame.
This is needed to prevent incorrect screenshots when drag
resizing. So for example, if the app was in split screen and
then the user dragged the divider to exit split screen, the WSA
would crop the area where the nav bar is and below. This would cause
the screenshot to get cropped, only showing the area that was last
computed by setSurfaceBoundariesLocked. This isn't necessary since
in drag resizing, the WSA hasn't redrawn yet so there's no need to
recalculate the crop for the area under the nav bar.

Bug: 76442549
Test: Enter split screen. Drag divider to exit. Ensure recents
screenshot is correct and not cropped.

Change-Id: If7f2639794a0c3a20353d40fe31ef9bb9a051e41

Former-commit-id: 9513662194e06817ea405255b6d322883785c0e6