Fix FingerprintDialogView when show is called before animation is complete
If handleShowDialog is called before the animation is finished, the window
will still be attached despite the mDialogShowing variable == false. We
need a new variable to track the dismiss animation state; if the dialog
is dismissing and handleShowDialog is called again, cancel the animation
and remove the window. Then show the dialog without an entry animation.
This lets us show the dialog without a "bouncy" animation when the device
rotates, etc..

Fixes: 77877888

Test: Manual test with modified settings APK
Change-Id: Iae878ca10099d9755c27284d6df04a73fd5a37f7

Former-commit-id: 87df068ac1184449879b595fe4986ff7e664aa16