Update nav buttons current view after inflates occur
Made sure that after nav buttons are inflated that the current view is
set immediately so that when using getCurrentView() you may not get an
older view (that could cause issues). Fixes the inability to quick scrub
because the getCurrentView() returned an old view that was detached from
parent and it's position was wrong not allowing home button to be
dragged.

Change-Id: Ia9ef61712298d0391b6479ee14d6410ca6d3f980
Fixes: 77909224
Test: wipe phone and try quick scrub

Former-commit-id: 1efc9c389676a49096a8755e14011fef5758333f