Update checks that AccessibilityNodeInfo is tested
We had a tests that detected when we added fields, but
there were existing fields that weren't tested, and a
couple that weren't even parceled. Updating the test
to make things more explicity, now that I've fixed
the cts test.

Also adding a similar check for boolean properties,
which were similarly lacking in coverage.

Bug: 77928402
Test: atest accessibility.AccessibilityNodeInfoTest
Change-Id: Ie8c52087343c3ccd2b8c86e5d7b9333b9294b138

Former-commit-id: c220ad6259adbe57f3e1697348011037719a061b