Fix layering of non-IME target windows in IME-target app in split-screen.
In split-screen mode we elevate child windows of the IME target with relative
layering to ensure they and the IME can exist above the docked divider while
the IME target itself still exists below. For behavior compatibility with O
we need to give this same treatment to all windows with the same token as the
IME target.

Bug: 70811741
Test: Manual. go/wm-smoke.
Change-Id: Ife174069ec2571c95d546981d196b7f519bb08ca

Former-commit-id: 4a1cdbd4e90a71290214f17b94ab65125e64d896