Have doze update the scrim controller rather than the scrim state.
Right now, when Doze wants to update the front alpha in AOD, it updates
the scrim state's front alpha value. Unfortunately, this is only looked
at when transitioning into the state, but entirely ignored once we're in
it. With this, we instead tell the scrim controller which tells the
scrim state but also updates the current front scrim alpha if we're
currently in the AOD state.

Bug: 78124574
Bug: 70939994
Test: atest ScrimControllerTest\#transitionToAod_withFrontAlphaUpdates
Change-Id: I09b33208bd3e3fd6d3829b33776364aa7b61a4f6

Former-commit-id: cae88753b0724d7f409146eb96fb96ab10970ac7