Don't start Window enter animation as a result of AWT becoming visible.
There are some cases where AWT becomes visible without an animation. It
will then start a WindowState enter animation if WS is not visible yet.
However, this is problematic since the WS may not be ready to become
visible yet. For example, relayout may not have been called yet to
create a surface and the size of the surface is not yet set. This causes
problems when determining the position of the leash when the animation
is created. Instead allow the WS relayout to handle starting the
enter animation so the WS is set properly.
As for the case where AWT is becoming not visible without an animation, but
WS is still visible, create an exit animation for the WS. This is
because there won't be a client to trigger this animation.

Change-Id: Ic9215a0c96f219f5bc87b16067755af582883a5d
Fixes: 77447982
Test: Issue from bug

Former-commit-id: e390cbd937fb7edbcba8abf8314f9c8074838684