Move status bar accessibility title to pane
The window title is new. It's actually confusing
TalkBack because announcing "Status Bar" when
the notification shade appears isn't helpful.

Since the window itself pretty much just holds
a bunch of stuff that is associated with the
top of the screen, we're actually removing
its title.

Re-adding the status bar title as a pane title
on the actual View that makes the bar.

Bug: 77984839
Test: TalkBack now announces Notification shade
properly rather than Status Bar.

Change-Id: I394009584b0a90a386dab5128654f943acecad47

Former-commit-id: 06c9e011d3cb7592c80ea4b30b34a6f69f459a3d