Fix unnecessarily extended View lifetime in IMMS
Currently InputMethodManagerService (IMMS) unnecessarily extends the
lifetime of View objects used to show IME switcher dialog until the
dialog is shown again, because it forgets to clear
IMMS#mSwitchingDialogTitleView with null even after the dialog is
closed.

As a result, roughly 1.5MB of RAM will be unnecessarily consumed by
the system_server once after the IME switcher dialog is shown until
the device is rebooted.

With this CL, IMMS#mSwitchingDialogTitleView will be cleared with null
when the dialog is closed.

Fixes: 78245299
Test: Manually made sure that IMMS does not hold strong references to
      View objects used for IME switcher dialog after the dialog is
      closed.
Change-Id: I9fee92bb2974edeceacc680686095df480efba29

Former-commit-id: 1fc7a1d4d34fbdb621e6339b4750ea152bd66f6f