Let SettingsLib not depend on InputMethodSettings
This is part of an effort to remove dependencies on InputMethodUtils
from Settings and other non-core Framework code.

Currently InputMethodSettingValuesWrapper depends on
InputMethodUtils.InputMethodSettings just to parse
Settings.Secure.ENABLED_INPUT_METHODS to extract enabled IMEs from the
result of InputMethodManager#getInputMethodList(), which can be
rewritten with InputMethodAndSubtypeUtil.

Also, seems that the dependency on
InputMethodUtils.InputMethodSettings confused the author and some
unnecessary synchronization blocks were introduced.

This CL completely removes the dependency on InputMethodSettings as
well as unnecessary synchronization blocks because methods of
InputMethodSettingValuesWrapper are expected to be called only from
the UI thread.

There should be no user-visible behavior change.

Bug: 77730201
Test: Manually verified that IME enabler still works as expected.
Change-Id: I0c52fb31ac2bb976dcd867f72992e260aee255b7

Former-commit-id: f367bc8fde3619a1db7d77ff9d41f3a7d8bc3eb4