Fixed CTS CtsUsageStatsTestCases test cases
We used the current RAT to load different carrier configuration to
determine whether the connection of a certain APN type is metered or
not. When the device is switched to IWLAN, the existing default
APN connection on cellular will be marked as unmetered based on
the current logic. This logic will be refactored when IWLAN full mode
is supported in Q.

For the short term workaround for P, we'll just mark default APN as
metered when the device is on IWLAN.

Test: CtsUsageStatsTestCases
Bug: 74540336, 77504652
Change-Id: Ib098e80d2bfd10e74c0b0dacf2cf2f217a66f7e4

Former-commit-id: 1e9793f926eec721c497d7d63513925628d4cceb